251101_시드임포터수정 후(빌드는 성공: 시드 재고 힌트 작업중)

+++ lib/src/screens/bom/order_shortage_result_screen.dart
@@ -264,7 +264,7 @@ class _LineVm {
   final String finishedItemId;
   final int orderQty;
   final int finishedStock;
-  final int finishedShortage;
+  final double finishedShortage;
   final List<_RowVm> semi;
   final List<_RowVm> raw;
   final List<_RowVm> sub;
